fix: moved getJobId and getAppScriptUrl to env.ts
