+++ src/__tests__/unit/analytics.spec.ts
@@ -1,11 +1,11 @@
 import { basename } from 'path';
 import { shareAnalyticsWithAppScript } from '../../analytics';
-import { getAppScriptUrl, getJobId } from '../../utils';
+import { getAppScriptUrl, getJobId } from '../../env';
 import { httpRequest } from '../../services/http.service';
 import logger from '../../../logger';
 import { filePath, fileName, fileContent } from '../../types/renameFiles';
 
-jest.mock('../../utils', () => ({
+jest.mock('../../env', () => ({
     getAppScriptUrl: jest.fn(),
     getJobId: jest.fn(),
 }));
